fix: FLOAT KESINLIGI — uc ayri sessiz hata, ust uste yiginmis
`float` calisma zamaninda double (backend->float_type = LLVMDoubleType,
VMValue payload 8 bayt). Ama uc yerde float32 varsayimi vardi ve ucu
birbirini gizliyordu.

1) ASTNode_C.value.float_value `float` idi (32-bit). Yani KAYNAKTAKI her
   float literali C-koprusunde float32'ye KIRPILIYORDU:
   `float pi = 3.141592653589793;` gercekte 3.1415927410125732 olarak
   derleniyordu. Uyari yok, hata yok. ~9 hane sessizce gidiyordu.

2) aot_format_float degeri once `(float)`e yuvarlayip en kisa float32
   gosterimini ariyordu. Yorumundaki gerekce — "the LLVM backend uses a
   32-bit float type" — YANLIS: `LLVMFloatTypeInContext` bu repoda hic
   olmadi (git log -S ile bakildi). Yani bicimleyici tuttugu sayidan
   BASKA bir sayi yaziyordu, ve tam da bu (1)'i gizliyordu: kirpilmis
   deger kirpilmis bicimle basilinca fark gorunmez.

3) `print` bu bicimleyiciyi HIC kullanmiyordu. Codegen aot_print_value
   cagiriyor, o da vm_print_value'ya (vm.cpp) gidiyor ve orasi duz
   `printf("%g")` — ALTI anlamli hane. Sonuc: `print(1000000.5)` ->
   "1e+06" (.5 tamamen kayboluyor), ama `print("" + x)` -> "1000000.5".
   Ayni deger, iki farkli sayi. runtime_bindings.cpp'deki dogru
   print_value/print_vm_value cifti AOT yolunda olu koddu (VM'den kalma).

Simdi ucu de double: 0.1+0.2 -> 0.30000000000000004,
3.141592653589793 tam basiliyor, print ile toString ayni sayiyi yaziyor.
Python/Go/Rust/JS ile ayni davranis. Siradan sayilarin bilimsele
kacmamasi (30.0 -> "30") korundu.

tests/float_format.test.tpr guncellendi: `toString(0.1+0.2) == "0.3"`
beklentisi float32 yuvarlamasini KODLUYORDU — yani testin kendisi hatayi
sabitliyordu. Basligindaki "Tulpar floats carry 32-bit precision" ifadesi
de yanlisti, duzeltildi.

ENJEKSIYON NOTU (onemli): (3)'u geri alinca float_precision paketi YESIL
kaldi — cunku assert'ler toString'den geciyor, print'ten degil. print'in
bicimleyicisi YALNIZ stdout karsilastiran sondada gorunuyor. Sondaya 6
print vakasi eklendi; simdi (3) enjekte edilince 7 sonda kirmiziya
donuyor. Testin olctugunu enjeksiyon soyledi, ben degil.

Sonda 48 -> 62 vaka (eleman bilesik atama + sarkan onbellek + print).

+++ src/parser/parser.cpp
@@ -1676,7 +1676,7 @@ static ASTNode_C* convert_ast_node(const ASTNode& node) {
             set_loc(out, n.loc);
         } else if constexpr (std::is_same_v<T, FloatLiteral>) {
             out->type = AST_FLOAT_LITERAL;
-            out->value.float_value = static_cast<float>(n.value);
+            out->value.float_value = n.value;   // double -> double, KIRPMA YOK
             set_loc(out, n.loc);
         } else if constexpr (std::is_same_v<T, StringLiteral>) {
             out->type = AST_STRING_LITERAL;
